Updated to version 1.3
Fixed an issue when deleting all rooms when a room had been selected, causing a crash when adding a new room.
Fixed potential crash when inserting items into the room array. Now this is treated the same as adding a room.
Fixed an issue with showing new room gizmo when ShowAllGizmo's is on.
Fixed an issue when manually deleting a Wall or Tile while using the tool.
Fixed issues when manually deleting a room while in Wall or Tile editmode.
Fixed cleanup on exiting the tool, so that relevant data is cleared.
Added syncing tool and scene selection of rooms in ManageRooms mode, allowing for easy deletion.
Added checks to trigger execution halt when internal state of tool is broken due to currently unsupported actions such as undoing/redoing room deletion/addition, to inform user to exit and reopen the tool to avoid crashing.